chỉnh sửa kiểu dữ liệu

+++ case_study_module_4/src/main/java/com/example/case_study_module_4/controller/UserController.java
@@ -21,8 +21,8 @@ public class UserController {
 
 
 
-    private Integer currentUserId() {
-        return 1;
+    private Long currentUserId() {
+        return 1L;
     }
 
     @GetMapping("/edit")
